test(inference_optimizer): hoist duplicated session_dir fixture into conftest
Eleven test modules each defined a byte-identical `session_dir` pytest
fixture (set USER_DATA_PATH -> make_session_dir() -> seed target-analysis
marker). Move it once into tests/conftest.py, where pytest auto-discovers
it, and delete the 11 local copies plus the now-dead make_session_dir
imports. Behavior is unchanged (fixture resolution is transparent); modules
may still shadow it locally if they need different seeding.

Net test reduction: -108 lines (measured via git diff --numstat: +16/-124).
Tests still pass (the one pre-existing test_resume failure is an unrelated
missing-Magpie-module env issue, present before this change).
